requirements.txt
bcrypt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
sqlalchemy
psycopg2-binary  # PostgreSQL driver